refactored component
